requirements_test.txt
asynctest==0.13.0
aioresponses==0.7.8
codecov==2.1.13
coverage==7.10.6
jsonpickle==4.1.1
mock-open==1.4.0
mypy==1.18.2
pylint==3.3.8
astroid==3.3.11
pipdeptree==2.26.1
pylint-strict-informational==0.1
pytest-aiohttp==1.1.0
pytest-asyncio==1.2.0
pytest-cov==7.0.0
pytest-test-groups==1.2.1
pytest-sugar==1.0.0
pytest-timeout==2.4.0
pytest-xdist==3.8.0
pytest==8.4.2
requests-mock==1.12.1
responses==0.25.8
stdlib-list==0.11.1
tqdm==4.67.1
pytest-homeassistant-custom-component==0.13.285
